updated readiness for test in milestone 1
