+++ object_detection/CMakeLists.txt
@@ -13,6 +13,7 @@ find_package(catkin REQUIRED COMPONENTS
   rospy
   std_msgs
   vortex_msgs
+  cv_msgs
 )
 
 ## System dependencies are found with CMake's conventions
