+++ object_detection/launch/object_detection_node.launch
@@ -1,16 +1,19 @@
 <?xml version="1.0"?>
 <launch>
-
+    <!-- Object detection node -->
     <node name="object_detection_node"    pkg="object_detection"     type="object_detection_node.py"    output="screen"/>
+    <node name="feature"    pkg="object_detection"     type="object_detection_node.py"    output="screen"/>
 
     <!-- <include file="$(find zed_wrapper)/launch/zed2.launch" /> -->
     <node pkg="tf2_ros" type="static_transform_publisher" name="detection" args="0 0 0 0 0 0 zed2_left_camera_frame detection" />
     <!-- <node pkg="tf2_ros" type="static_transform_publisher" name="detection" args="0 0 0 0 0 0 map detection" /> -->
 
     <!-- <node pkg="tf2_ros" type="static_transform_publisher" name="detection" args="0 0 0 0 0 0 1 map detection" /> -->
-
     <node name="test"    pkg="landmarks"     type="landmarks"    output="screen"/>
 
 
-
+    <!-- Feature detection node -->
+    <node name="republish" type="republish" pkg="image_transport" output="screen" args="compressed in:=/zed2/zed_node/rgb/image_rect_color raw out:=/zed2/zed_node/rgb/image_rect_color" />
+    <node name="feature_detection_node"    pkg="feature_detection"     type="feature_detection_node.py"    output="screen"/>
+    <node name="CVOD"    pkg="feature_detection"     type="dynam_recfg.py"    output="screen"/>
 </launch>
