+++ object_detection/package.xml
@@ -54,16 +54,19 @@
   <build_depend>rospy</build_depend>
   <build_depend>std_msgs</build_depend>
   <build_depend>vortex_msgs</build_depend>
+  <build_depend>cv_msgs</build_depend>
   <build_export_depend>darknet_ros_msgs</build_export_depend>
   <build_export_depend>roscpp</build_export_depend>
   <build_export_depend>rospy</build_export_depend>
   <build_export_depend>std_msgs</build_export_depend>
   <build_export_depend>vortex_msgs</build_export_depend>
+  <build_export_depend>cv_msgs</build_export_depend>
   <exec_depend>darknet_ros_msgs</exec_depend>
   <exec_depend>roscpp</exec_depend>
   <exec_depend>rospy</exec_depend>
   <exec_depend>std_msgs</exec_depend>
   <exec_depend>vortex_msgs</exec_depend>
+  <exec_depend>cv_msgs</exec_depend>
 
 
   <!-- The export tag contains other, unspecified, tags -->
